Adding ability to do unsafe operations
"Unsafe", aka changing my interview schedule if it is an earlier time
It works though. With this script, it helped me got an interview slot
happening in three days, as opposed to the next available interview
slot, which four months out.

Here's my time schedule in case anybody is in the same boat as me:
Submitted my NEXUS application on 6/6/13.
Received my conditional approval on 8/8/13.
Next available interview slot in Seattle Urban center is 12/20/13.
I made this script over the weekend, and it successfully changed
and pushed my date to 8/15/13, because someone canceled.

Success? I think so