SRM 154 DIV 1 - 134.39 pts for 350pt problem
The tl;dr version of the problem is, find the substring of code in
keyPresses, but the tricky part is, consequetive chars in the code has
to be greater than the ones in keyPresses.

So my approach was to let regex to the work. I set a lower bound of
repeatability, but no upper bound. Took me a while to get it right
though.